% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot_Type1.R
\name{Boxplot_TypeI}
\alias{Boxplot_TypeI}
\title{Boxplot_TypeI creates a boxplot of the type I error results per method}
\usage{
Boxplot_TypeI(
  path,
  method = NULL,
  alpha = 0.05,
  tol = 0.02,
  panel = "",
  ylim = c(0, 0.13),
  N = 10000
)
}
\arguments{
\item{path}{to the NeutraliseFiles}

\item{method}{is a specified method (character string) or is NULL (default). When =NULL, a boxplot of all methods is made.}

\item{alpha}{is the significance level, default=0.05}

\item{panel}{is a paramter to specifiy the boxplot based on 'n'=samplesize or 'distribution'= data generation method}

\item{ylim}{are the limits of the y-axis of the boxplot}

\item{N}{is the amount of simulations the results come from (default = 10 000)}
}
\value{
a list with 3 objects: 1) Graph = ggplot2 boxplot 2) text object that defines the results 3) a dataframe with the results the graph is based on
}
\description{
Boxplot_TypeI creates a boxplot of the type I error results per method
}
