% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutralise.R
\name{Neutralise}
\alias{Neutralise}
\title{Neutralise Main function}
\usage{
Neutralise(
  path,
  Test = NULL,
  Data.Generator = NULL,
  settings = NULL,
  B = 10000,
  N = 1000
)
}
\arguments{
\item{path}{is the directory path of neutralizefiles}

\item{Test, }{run simulations for this test}

\item{Data.Generator, }{generates data}
}
\value{
results
}
\description{
Neutralise Main function
}
\examples{

}
