% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power_curve.R
\name{Power_curve}
\alias{Power_curve}
\title{Power_curve creates a curve plot of the power in function of delta}
\usage{
Power_curve(
  path,
  methods = NULL,
  alpha = 0.05,
  par.fix = NULL,
  data = NULL,
  col = "black",
  CI = FALSE,
  N = 10000
)
}
\arguments{
\item{path}{is the path to NeutraliseFiles}

\item{methods}{is a character string that specifies a method (default =NULL, when =NULL a power curve is made for all methods seperatly)}

\item{alpha}{is the significance level (default = 0.05)}

\item{par.fix}{is a data frame that specifies a specific setting parameter (default =NULL)}

\item{data}{is a character string that defines a data generation method (default = NULL, its necessary to specify this as the power-cruve will be unclear)}

\item{CI}{is a logical variable, when =TRUE the CI for the observations wil be plotted}

\item{N}{is the amount of simulations (default = 10000)}
}
\value{
Power curve
}
\description{
Power_curve creates a curve plot of the power in function of delta
}
