% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power_QQ.R
\name{Power_QQ}
\alias{Power_QQ}
\title{Power_QQ creates a plot that compare the power results of 2 specified methods}
\usage{
Power_QQ(
  path,
  method1,
  method2,
  alpha = 0.05,
  par.fix = NULL,
  data = NULL,
  col = "black",
  group = FALSE,
  N = 10000
)
}
\arguments{
\item{path}{is the path to NeutraliseFiles}

\item{method1}{is a character string that specifies a method}

\item{method2}{is a character string that specifies a second method}

\item{alpha}{is the significance level (default = 0.05)}

\item{par.fix}{is a data frame that specifies a specific setting parameter (default =NULL)}

\item{data}{is a character string that defines a data generation method (default = NULL)}

\item{group}{is a logical variable that indicates to color the observations based on data generation method}

\item{N}{is the amount of simulations (default = 10000)}
}
\value{
a list object with 6 elements. 1) amount of times method1 has a higher power than method2 2) results1 - results data frame of method 1 3) results2 - results data frame of method 2 4) total - data frame of all the results for the graph 5) graph 6) text string that summarizes the results
}
\description{
Power_QQ creates a plot that compare the power results of 2 specified methods
}
